+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Pulse: refill all available worker slots immediately after worker exit (#29448)
+
 ## [3.32.219] - 2026-08-03
 
 ### Added